[refactor] hw_L44: the file 'index.js' is divided into categories 'app.js' and 'functions.js' according to comments from gitHub
